# tools/Swagger/swagger_V0.9.5.26.json
{
...
  "swagger": "2.0",
  "info": {
    "title": "OrderSolution HTTP API",
    "version": "0.0.0"
  },
  "host": "10.0.0.136:12346",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json-patch+json",
    "application/json",
    "text/json",
    "application/*+json"
  ],
  "paths": {
...
    "/api/v2/actions/OrderLines/Void/{orderLineId}": {
      "post": {
        "tags": [
          "actions/orderlines"
        ],
        "summary": "Void an orderline or a given quantity of it.",
        "operationId": "OrderLines_Void",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "type": "string",
            "name": "orderLineId",
            "in": "path",
            "required": true,
            "description": "The order line id to void / to reduce the quantity of.",
            "x-nullable": true
          },
          {
            "name": "data",
            "in": "body",
            "required": true,
            "description": "Details about the void operation to be executed on this orderline. \n            If null the whole orderline will be voided.",
            "schema": {
              "$ref": "#/definitions/OrderLineVoid"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/OrderLineVoidResult"
            }
          },
          "204": {
            "description": ""
          }
        }
      }
    },
    "/api/v2/actions/OrderLines/Add/{subTableId}": {
      "post": {
        "tags": [
          "actions/orderlines"
        ],
        "operationId": "OrderLines_Add",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "type": "string",
            "name": "subTableId",
            "in": "path",
            "required": true,
            "x-nullable": true
          },
          {
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/OrderLineAdd"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "201": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/OrderLineResult"
            }
          }
        }
      }
    },
    "/api/v2/actions/OrderLines/ModifyUncommitted/{orderLineId}": {
      "post": {
        "tags": [
          "actions/orderlines"
        ],
        "summary": "Modifies (= replaces) an uncommited orderline.\nIMPORTANT: Any existing modifiers of the orderline are replaced with the new modifiers. \nFor example if an orderline \"1x Coke with ice\" is modified to be \"1x Coke with Citron\" the ice is also gone\nso the operation works as a replacement and not as a patch of the existing data.",
        "operationId": "OrderLines_ModifyUncommitted",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "type": "string",
            "name": "orderLineId",
            "in": "path",
            "required": true,
            "description": "The orderline to be modified / replaced",
            "x-nullable": true
          },
          {
            "name": "data",
            "in": "body",
            "required": true,
            "description": "The new data to be used for this order line (same as for the add request)",
            "schema": {
              "$ref": "#/definitions/OrderLineModify"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "201": {
            "x-nullable": true,
            "description": "OrderLineResult informing about the new ids / prices of the modified orderline.",
            "schema": {
              "$ref": "#/definitions/OrderLineResult"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "OrderLineVoidResult": {
      "type": "object",
      "description": "Provides information about the new status of the orderline after void processing.",
      "additionalProperties": false,
      "required": [
        "singlePrice",
        "quantity"
      ],
      "properties": {
        "orderLineId": {
          "type": "string"
        },
        "singlePrice": {
          "type": "integer",
          "format": "int32"
        },
        "quantity": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "OrderLineVoid": {
      "type": "object",
      "additionalProperties": false,
      "properties": {
        "quantity": {
          "type": "integer",
          "description": "Defines the quantity that is voided from an existing orderline.\nIf the quanity is 0 nothing is changed.\nIf the quantity is &gt;= the quantity of the orderline the whole orderline will be voided.",
          "format": "int32"
        },
        "cancellationReasonId": {
          "type": "string",
          "description": "Defines the cancellation reasion that must be provided if a already commited orderline should\nbe voided."
        }
      }
    },
    "OrderLineResult": {
      "type": "object",
      "additionalProperties": false,
      "required": [
        "singlePrice"
      ],
      "properties": {
        "id": {
          "type": "string"
        },
        "singlePrice": {
          "type": "integer",
          "format": "int32"
        },
        "modifiers": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OrderLineResultModifier"
          }
        }
      }
    },
    "OrderLineResultModifier": {
      "type": "object",
      "additionalProperties": false,
      "properties": {
        "id": {
          "type": "string"
        },
        "choices": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OrderLineResultModifierChoice"
          }
        }
      }
    },
    "OrderLineResultModifierChoice": {
      "type": "object",
      "additionalProperties": false,
      "properties": {
        "modifierChoiceId": {
          "type": "string"
        },
        "choicePrice": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "OrderLineAdd": {
      "type": "object",
      "additionalProperties": false,
      "properties": {
        "quantity": {
          "type": "integer",
          "description": "The quantity for this order line. When adding a new order line this value must be set.",
          "format": "int32"
        },
        "enteredPrice": {
          "type": "integer",
          "description": "Defines a manual price for this order line.\nNOTE: You must not provide this parameter if you want the price to be calculated by the POS.",
          "format": "int32"
        },
        "modifiers": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OrderLineModifier2"
          }
        },
        "articleId": {
          "type": "string"
        }
      }
    },
    "OrderLineModifier2": {
      "type": "object",
      "additionalProperties": false,
      "properties": {
        "modifierGroupId": {
          "type": "string"
        },
        "textInput": {
          "type": "string"
        },
        "faxInputID": {
          "type": "string",
          "description": "The ID of the FaxInput image. Can be a number, url, ...\nThe format if this ID is defined by the POS storage behavior.\nTODO: When the FaxInput storage API is defined we may have to update this."
        },
        "choices": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OrderLineModifierChoice2"
          }
        }
      }
    },
    "OrderLineModifierChoice2": {
      "type": "object",
      "additionalProperties": false,
      "properties": {
        "modifierChoiceId": {
          "type": "string"
        },
        "amount": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "OrderLineModify": {
      "type": "object",
      "additionalProperties": false,
      "properties": {
        "quantity": {
          "type": "integer",
          "description": "The quantity for this order line. When adding a new order line this value must be set.",
          "format": "int32"
        },
        "enteredPrice": {
          "type": "integer",
          "description": "Defines a manual price for this order line.\nNOTE: You must not provide this parameter if you want the price to be calculated by the POS.",
          "format": "int32"
        },
        "modifiers": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OrderLineModifier2"
          }
        }
      }
    },
...
  }
}